+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## DEVELOPMENT
+## CURRENT
 
 ## 0.13.2
 
@@ -13,8 +13,7 @@ Tweaked
 Fixed
 - Fixed the `removeTentAndResearcher` gamemaster event removing the first donkey it found, even if not owned by the researcher
 - Fixed creepers not exploding inside researcher's tents even when no researcher is present
-
-## CURRENT
+- Fixed the logs on Neoforge having two "Ruins of Growsseth" prefixes
 
 ## 0.13.1